inital debian packaging env and stock boilerplate
This is an initial commit that contains some basic boilerplate for the
current thought of what a packaging repo should look like.

The debian directory contains a Vagrantfile that is currently
provisioned using the shell provisioner and some boilerplate env files
that are sourced at login by the vagrant user.

This should not be relied upon at this point and is merely being used to
get the package built as quick as possible. In the very near future
Ansible will be brought in so that multiple platforms can be configured
with greater ease.

Feel free to open an issue with any comments or concerns regarding this
process.