add debian role
This role will be the foundation for working with Debian. Currently it
will include the following functions:

- updating the currently installed packages
- setting the upstream sources
- installing packages necessary for building the software
- installing packages necessary for packaging the software
- templating debian specific files needed for packaging

There are no plans at this time to perform version updates. When new
versions of Debian are released then we will just pull a new box. The
sources are being managed solely for the purpose of testing the software
and builds against the lastest available as well as stable to ensure
max compatability.